User controller implemented, code optimized.
