+++ routes/transaction.route.js
@@ -1,4 +1,3 @@
-require("dotenv").config();
 const express = require("express");
 const router = express.Router();
 const authValidate = require("../middleware/auth");
